pyproject.toml
[tool.poetry]
name = "jira-analog"
version = "0.1.0"
description = "Jira Analog app - university project"
authors = ["Ivan Yolgin <yolginivan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9.10"
gunicorn = "^20.1.0"
uvicorn = "^0.17.6"
python-dotenv = "^0.19.2"
fastapi-mail = "^1.0.4"
PyJWT = "^2.3.0"
aerich = "^0.6.2"
asyncpg = "^0.25.0"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
Authlib = "^1.0.0"
python-slugify = "^6.1.1"
fastapi-pagination = "^0.9.1"


[tool.poetry.dev-dependencies]
pylint = "^2.12.2"
pytest = "^7.1.0"
pre-commit = "^2.17.0"
black = "^22.1.0"
flake8 = "^4.0.1"


[tool.aerich]
tortoise_orm = "backend.config.tortoise_conf.TORTOISE_ORM"
location = "./backend/migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.idea
  | \.pytest_cache
  | \migrations
  | _build
  | buck-out
  | build
  | dist
)/
'''